Rename external_sign to cose_sign::sign_claim, add documentation, and make this function public
