+++ sdk/src/lib.rs
@@ -80,9 +80,6 @@ mod cose_validator;
 mod error;
 pub use error::{Error, Result};
 
-mod external_signer;
-pub use external_signer::external_sign;
-
 mod ingredient;
 pub use ingredient::{Ingredient, IngredientOptions};
 pub mod jumbf_io;
@@ -117,8 +114,10 @@ pub(crate) mod assertion;
 pub(crate) mod asset_handlers;
 pub(crate) mod asset_io;
 pub(crate) mod claim;
+
 #[cfg(feature = "file_io")]
-pub(crate) mod cose_sign;
+pub mod cose_sign;
+
 #[cfg(all(feature = "xmp_write", feature = "file_io"))]
 pub(crate) mod embedded_xmp;
 pub(crate) mod hashed_uri;
